added clean false to _init_repo.sh

+++ _init_repo.sh
@@ -12,23 +12,25 @@ function init_repo() {
     git config --global user.name "Git Master"
   fi
 
-  REPO_NAME=$1
+  _REPO_NAME=$1
   SECONDARY_REPO=$2
 
   GIT_REPOS=.git-repos
 
-  echo "Cleaning $(pwd)/workspace/*"
+  if [ -z "$CLEAN" ]; then
+    echo "Cleaning $(pwd)/workspace/*"
 
-  rm -rf "$(pwd)/workspace"
-  mkdir -p "$(pwd)/workspace"
-  rm -rf $GIT_REPOS
+    rm -rf "$(pwd)/workspace"
+    mkdir -p "$(pwd)/workspace"
+    rm -rf $GIT_REPOS
+  fi
 
-  mkdir -p $GIT_REPOS/$REPO_NAME.git
-  git init --initial-branch "main" --bare $GIT_REPOS/$REPO_NAME.git
+  mkdir -p $GIT_REPOS/$_REPO_NAME.git
+  git init --initial-branch "main" --bare $GIT_REPOS/$_REPO_NAME.git
 
-  git clone file://$(pwd)/$GIT_REPOS/$REPO_NAME.git "$(pwd)/workspace/$REPO_NAME"
+  git clone file://$(pwd)/$GIT_REPOS/$_REPO_NAME.git "$(pwd)/workspace/$_REPO_NAME"
 
-  cd "workspace/$REPO_NAME"
+  cd "workspace/$_REPO_NAME"
   git config init.defaultBranch "main"
   git commit -m "init" --allow-empty
   git push origin main
@@ -37,8 +39,8 @@ function init_repo() {
 
   if [ ! -z $SECONDARY_REPO ];
   then
-    git clone file:////$(pwd)/$GIT_REPOS/$REPO_NAME.git "$(pwd)/workspace/${REPO_NAME}-2"
-    cd "workspace/${REPO_NAME}-2"
+    git clone file:////$(pwd)/$GIT_REPOS/$_REPO_NAME.git "$(pwd)/workspace/${_REPO_NAME}-2"
+    cd "workspace/${_REPO_NAME}-2"
     git config init.defaultBranch "main"
     aliases
     cd -
